pyproject.toml
[project]
name = "img2svg-bench"
version = "0.1.0"
description = "A local-first benchmark harness for comparing image-to-SVG tools and presets."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pillow>=12.1.1",
  "vtracer>=0.6.15",
]

[tool.uv]
package = false
